extract context menu into dedicated component
